// backend/proto/randomPackage/User.ts
// Original file: E:/Programming projects/Full stack Web dev/React-grpc chatapp/backend/proto/random.proto


export interface User {
  'id'?: (number);
  'name'?: (string);
  'status'?: (string);
  'avatarUrl'?: (string);
}

export interface User__Output {
  'id'?: (number);
  'name'?: (string);
  'status'?: (string);
  'avatarUrl'?: (string);
}