// backend/proto/randomPackage/StreamMessage.ts
// Original file: E:/Programming projects/Full stack Web dev/React-grpc chatapp/backend/proto/random.proto


export interface StreamMessage {
  'userId'?: (number);
  'userAvatar'?: (string);
  'message'?: (string);
  'userName'?: (string);
}

export interface StreamMessage__Output {
  'userId'?: (number);
  'userAvatar'?: (string);
  'message'?: (string);
  'userName'?: (string);
}